requirements.txt
altair>=4.0.0
google-generativeai
matplotlib>=3.2.2
numpy
opencv-python-headless
pandas
Pillow>=9.0.0
python-dotenv>=0.19.0
PyYAML>=5.3.1
requests>=2.25.0
scipy
seaborn
streamlit
tensorboard
thop
torch
torchvision
tqdm>=4.41.0
scikit-learn
transformers
huggingface-hub
retina-face
openpyxl
